Fix SPI type mismatch: IoHwAb uint8* vs Spi.h uint16*
IoHwAb declared Spi_ReadIB/WriteIB with uint8* while Spi.h uses
uint16*. Spi_ReadIB copies 4 uint16 words (8 bytes) into a 2-byte
uint8 buffer — stack buffer overflow corrupting pedal angle data,
causing stuck detection and RUN→DEGRADED on the SIL demo.

Fix: align IoHwAb externs to uint16*, use uint16[4] buffers matching
SPI_IB_SIZE, read angle directly from rx_data[0] & 0x3FFF. Update
Spi_Posix stub to write angle as uint16 word (no byte-swap). Update
test mocks to match uint16* signatures.

+++ firmware/shared/bsw/ecual/IoHwAb.h
@@ -79,9 +79,9 @@ typedef struct {
 
 /* ---- External MCAL Dependencies ---- */
 
-/* SPI driver */
-extern Std_ReturnType Spi_WriteIB(uint8 Channel, const uint8* DataBufferPtr);
-extern Std_ReturnType Spi_ReadIB(uint8 Channel, uint8* DataBufferPtr);
+/* SPI driver (signatures must match Spi.h — uint16 word buffers) */
+extern Std_ReturnType Spi_WriteIB(uint8 Channel, const uint16* DataBufferPtr);
+extern Std_ReturnType Spi_ReadIB(uint8 Channel, uint16* DataBufferPtr);
 extern Std_ReturnType Spi_SyncTransmit(uint8 Sequence);
 
 /* ADC driver */
